2021.4.16 tsl commit 18:03 在促销服务加缓存

+++ car-marketing-server/car-marketing-model/src/main/java/com/woniu/car/marketing/model/dtoVo/GetCouponAllDtoVo.java
@@ -17,7 +17,7 @@ public class GetCouponAllDtoVo {
 
     private BigDecimal couponMoney;//优惠券面额(元)
 
-    private Integer couponCondition;//使用门槛（满多少可用）
+    private BigDecimal couponCondition;//使用门槛（满多少可用）
 
     private Integer couponGoods;//发行来源(0平台-其他对应门店id)
